Put all code from chargeIngestion into a handler
