+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.FunctionHost/Charges/ChargeIngestion.cs
@@ -12,52 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using System.ComponentModel;
 using System.Threading.Tasks;
-using Energinet.DataHub.Core.App.Common.Abstractions.Actor;
-using Energinet.DataHub.Core.App.FunctionApp.Middleware.CorrelationId;
-using Energinet.DataHub.Core.Messaging.Transport.SchemaValidation;
-using GreenEnergyHub.Charges.Application.Charges.Handlers.ChargeInformation;
-using GreenEnergyHub.Charges.Application.Charges.Handlers.ChargePrice;
-using GreenEnergyHub.Charges.Domain.Dtos.ChargeInformationCommands;
-using GreenEnergyHub.Charges.Domain.Dtos.ChargePriceCommands;
-using GreenEnergyHub.Charges.Domain.Dtos.Messages.Command;
-using GreenEnergyHub.Charges.Infrastructure.CimDeserialization.MarketDocument;
-using GreenEnergyHub.Charges.Infrastructure.Core.Function;
-using GreenEnergyHub.Charges.Infrastructure.Core.MessagingExtensions;
+using GreenEnergyHub.Charges.FunctionHost.Charges.Handlers;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
-using Microsoft.Extensions.Logging;
 
 namespace GreenEnergyHub.Charges.FunctionHost.Charges
 {
     public class ChargeIngestion
     {
-        private readonly ILogger _logger;
-        private readonly ICorrelationContext _correlationContext;
-        private readonly IChargeInformationCommandBundleHandler _chargeInformationCommandBundleHandler;
-        private readonly IChargePriceCommandBundleHandler _chargePriceCommandBundleHandler;
-        private readonly IHttpResponseBuilder _httpResponseBuilder;
-        private readonly ValidatingMessageExtractor<ChargeCommandBundle> _messageExtractor;
-        private readonly IActorContext _actorContext;
+        private readonly IChargeIngestionHandler _chargeIngestionHandler;
 
-        public ChargeIngestion(
-            ILoggerFactory loggerFactory,
-            ICorrelationContext correlationContext,
-            IChargeInformationCommandBundleHandler chargeInformationCommandBundleHandler,
-            IChargePriceCommandBundleHandler chargePriceCommandBundleHandler,
-            IHttpResponseBuilder httpResponseBuilder,
-            ValidatingMessageExtractor<ChargeCommandBundle> messageExtractor,
-            IActorContext actorContext)
+        public ChargeIngestion(IChargeIngestionHandler chargeIngestionHandler)
         {
-            _logger = loggerFactory.CreateLogger(nameof(ChargeIngestion));
-            _correlationContext = correlationContext;
-            _chargeInformationCommandBundleHandler = chargeInformationCommandBundleHandler;
-            _chargePriceCommandBundleHandler = chargePriceCommandBundleHandler;
-            _httpResponseBuilder = httpResponseBuilder;
-            _messageExtractor = messageExtractor;
-            _actorContext = actorContext;
+            _chargeIngestionHandler = chargeIngestionHandler;
         }
 
         [Function(IngestionFunctionNames.ChargeIngestion)]
@@ -65,76 +33,7 @@ namespace GreenEnergyHub.Charges.FunctionHost.Charges
             [HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = null)]
             HttpRequestData request)
         {
-            try
-            {
-                SchemaValidatedInboundMessage<ChargeCommandBundle> inboundMessage;
-                try
-                {
-                    inboundMessage = await ValidateMessageAsync(request).ConfigureAwait(false);
-                }
-                catch (Exception exception) when (exception is InvalidXmlValueException or InvalidEnumArgumentException)
-                {
-                    _logger.LogError(
-                        exception,
-                        "Unable to parse request with correlation id: {CorrelationId}",
-                        _correlationContext.Id);
-
-                    return _httpResponseBuilder.CreateBadRequestB2BResponse(
-                        request, B2BErrorCode.SyntaxValidationErrorMessage, exception.Message);
-                }
-
-                if (inboundMessage.HasErrors)
-                {
-                    return await _httpResponseBuilder
-                        .CreateBadRequestResponseAsync(request, inboundMessage.SchemaValidationError)
-                        .ConfigureAwait(false);
-                }
-
-                if (AuthenticatedMatchesSenderId(inboundMessage) == false)
-                {
-                    return _httpResponseBuilder.CreateBadRequestB2BResponse(
-                        request, B2BErrorCode.ActorIsNotWhoTheyClaimToBeErrorMessage, string.Empty);
-                }
-
-                var bundle = inboundMessage.ValidatedMessage;
-                switch (bundle)
-                {
-                    case ChargeInformationCommandBundle commandBundle:
-                        ChargeCommandNullChecker.ThrowExceptionIfRequiredPropertyIsNull(commandBundle);
-                        await _chargeInformationCommandBundleHandler.HandleAsync(commandBundle).ConfigureAwait(false);
-                        break;
-                    case ChargePriceCommandBundle commandBundle:
-                        await _chargePriceCommandBundleHandler.HandleAsync(commandBundle).ConfigureAwait(false);
-                        break;
-                }
-
-                return _httpResponseBuilder.CreateAcceptedResponse(request);
-            }
-            catch (SchemaValidationException exception)
-            {
-                _logger.LogError(
-                    exception,
-                    "Unable to schema validate request with correlation id: {CorrelationId}",
-                    _correlationContext.Id);
-                return await _httpResponseBuilder
-                    .CreateBadRequestResponseAsync(request, exception.SchemaValidationError)
-                    .ConfigureAwait(false);
-            }
-        }
-
-        private bool AuthenticatedMatchesSenderId(SchemaValidatedInboundMessage<ChargeCommandBundle> inboundMessage)
-        {
-            var authorizedActor = _actorContext.CurrentActor;
-            var senderId = inboundMessage.ValidatedMessage?.Document.Sender.MarketParticipantId;
-
-            return authorizedActor != null && senderId == authorizedActor.Identifier;
-        }
-
-        private async Task<SchemaValidatedInboundMessage<ChargeCommandBundle>> ValidateMessageAsync(HttpRequestData req)
-        {
-            return (SchemaValidatedInboundMessage<ChargeCommandBundle>)await _messageExtractor
-                .ExtractAsync(req.Body)
-                .ConfigureAwait(false);
+            return await _chargeIngestionHandler.HandleAsync(request).ConfigureAwait(false);
         }
     }
 }
